Introduce NIOFileUtil to remove code duplication.

+++ src/main/java/org/seqdoop/hadoop_bam/util/SAMFileMerger.java
@@ -1,6 +1,5 @@
 package org.seqdoop.hadoop_bam.util;
 
-import com.google.common.annotations.VisibleForTesting;
 import com.google.common.io.CountingOutputStream;
 import htsjdk.samtools.SAMFileHeader;
 import htsjdk.samtools.cram.build.CramIO;
@@ -10,24 +9,21 @@ import htsjdk.samtools.util.BlockCompressedStreamConstants;
 import java.io.IOException;
 import java.io.InputStream;
 import java.io.OutputStream;
-import java.net.URI;
-import java.nio.file.FileVisitResult;
 import java.nio.file.Files;
 import java.nio.file.NoSuchFileException;
 import java.nio.file.Path;
-import java.nio.file.PathMatcher;
-import java.nio.file.Paths;
-import java.nio.file.SimpleFileVisitor;
-import java.nio.file.attribute.BasicFileAttributes;
 import java.util.ArrayList;
-import java.util.Collections;
 import java.util.LinkedList;
 import java.util.List;
-import java.util.stream.Collectors;
 import org.seqdoop.hadoop_bam.KeyIgnoringAnySAMOutputFormat;
 import org.seqdoop.hadoop_bam.SAMFormat;
 import org.seqdoop.hadoop_bam.SplittingBAMIndex;
 import org.seqdoop.hadoop_bam.SplittingBAMIndexer;
+
+import static org.seqdoop.hadoop_bam.util.NIOFileUtil.asPath;
+import static org.seqdoop.hadoop_bam.util.NIOFileUtil.deleteRecursive;
+import static org.seqdoop.hadoop_bam.util.NIOFileUtil.getFilesMatching;
+import static org.seqdoop.hadoop_bam.util.NIOFileUtil.mergeInto;
 
 /**
  * Merges headerless BAM or CRAM files produced by {@link KeyIgnoringAnySAMOutputFormat}
@@ -63,6 +59,14 @@ public class SAMFileMerger {
           "path: " + partPath);
     }
 
+    List<Path> parts = getFilesMatching(partPath,
+        "glob:**/part-[mr]-[0-9][0-9][0-9][0-9][0-9]*",
+        SplittingBAMIndexer.OUTPUT_FILE_EXTENSION);
+    if (parts.isEmpty()) {
+      throw new IllegalArgumentException("Could not write bam file because no part " +
+          "files were found in " + partPath);
+    }
+
     Files.deleteIfExists(outputPath);
 
     long headerLength;
@@ -72,7 +76,7 @@ public class SAMFileMerger {
         new SAMOutputPreparer().prepareForRecords(out, samOutputFormat, header); // write the header
       }
       headerLength = out.getCount();
-      mergeInto(out, partPath);
+      mergeInto(parts, out);
       writeTerminatorBlock(out, samOutputFormat);
     }
     long fileLength = Files.size(outputPath);
@@ -88,51 +92,6 @@ public class SAMFileMerger {
     deleteRecursive(partPath);
   }
 
-  private static Path asPath(String path) {
-    URI uri = URI.create(path);
-    return uri.getScheme() == null ? Paths.get(path) : Paths.get(uri);
-  }
-
-  private static void deleteRecursive(Path directory) throws IOException {
-    Files.walkFileTree(directory, new SimpleFileVisitor<Path>() {
-      @Override
-      public FileVisitResult visitFile(Path file, BasicFileAttributes attrs) throws IOException {
-        Files.delete(file);
-        return FileVisitResult.CONTINUE;
-      }
-      @Override
-      public FileVisitResult postVisitDirectory(Path dir, IOException exc) throws IOException {
-        Files.delete(dir);
-        return FileVisitResult.CONTINUE;
-      }
-    });
-  }
-
-  private static void mergeInto(OutputStream out, Path directory) throws IOException {
-    final List<Path> parts = getFragments(directory);
-    if (parts.isEmpty()) {
-      throw new IllegalArgumentException("Could not write bam file because no part " +
-          "files were found in " + directory);
-    }
-    for (final Path part : parts) {
-      Files.copy(part, out);
-    }
-    for (final Path part : parts) {
-      Files.delete(part);
-    }
-  }
-
-  @VisibleForTesting
-  static List<Path> getFragments(final Path directory) throws IOException {
-    PathMatcher matcher = directory.getFileSystem().getPathMatcher("glob:**/part-[mr]-[0-9][0-9][0-9][0-9][0-9]*");
-    List<Path> parts = Files.walk(directory)
-        .filter(matcher::matches)
-        .filter(path -> !path.toString().endsWith(SplittingBAMIndexer.OUTPUT_FILE_EXTENSION))
-        .collect(Collectors.toList());
-    Collections.sort(parts);
-    return parts;
-  }
-
   //Terminate the aggregated output stream with an appropriate SAMOutputFormat-dependent terminator block
   private static void writeTerminatorBlock(final OutputStream out, final SAMFormat samOutputFormat) throws IOException {
     if (SAMFormat.CRAM == samOutputFormat) {
@@ -144,8 +103,9 @@ public class SAMFileMerger {
 
   static void mergeSplittingBaiFiles(OutputStream out, Path directory, long headerLength,
       long fileLength) throws IOException {
-    final List<Path> parts = getSplittingBaiFiles(directory);
-
+    final List<Path> parts = getFilesMatching(directory,
+        "glob:**/part-[mr]-[0-9][0-9][0-9][0-9][0-9]*" +
+            SplittingBAMIndexer.OUTPUT_FILE_EXTENSION, null);
     if (parts.isEmpty()) {
       return; // nothing to merge
     }
@@ -186,13 +146,5 @@ public class SAMFileMerger {
     long blockAddress = BlockCompressedFilePointerUtil.getBlockAddress(virtualFilePointer);
     int blockOffset = BlockCompressedFilePointerUtil.getBlockOffset(virtualFilePointer);
     return (blockAddress + offset) << 16 | (long) blockOffset;
-  }
-
-  static List<Path> getSplittingBaiFiles(final Path directory) throws IOException {
-    PathMatcher matcher = directory.getFileSystem()
-        .getPathMatcher("glob:**/part-[mr]-[0-9][0-9][0-9][0-9][0-9]*" + SplittingBAMIndexer.OUTPUT_FILE_EXTENSION);
-    List<Path> parts = Files.walk(directory).filter(matcher::matches).collect(Collectors.toList());
-    Collections.sort(parts);
-    return parts;
   }
 }
